+++ src/dec/vp8i.h
@@ -275,7 +275,7 @@ struct VP8Decoder {
 
   // extensions
   const uint8_t* alpha_data_;   // compressed alpha data (if present)
-  uint32_t alpha_data_size_;
+  size_t alpha_data_size_;
   uint8_t* alpha_plane_;        // output
 
   int layer_colorspace_;
